Move PoseStreamer files
